Test/123 (#26)
* Update Go dependencies and replace logger

Refactor dependencies in go.mod and go.sum files, replacing 'github.com/google/logger' package with 'github.com/jinzhu/gorm' and 'github.com/lib/pq' packages to utilize ORM and PostgreSQL capabilities in the project. Also, some indirect dependencies were updated due to compatibility issues.

* Switched SQLite to Postgres and updated dependencies

The database driver was switched from SQLite to Postgres to better handle production requirements and concurrency. Additionally, a number of dependencies were updated to their latest versions to keep the project updated and secure. Various indirect dependencies were also added for using with the new Postgres database driver. Make sure to update your .env file to specify the new `DbType` and `DatabaseURL`.

* Add new Account model for PostgreSQL migration

A new Account model was added in the application domain to facilitate the migration from SQLite to PostgreSQL. The new struct is adjusted to match the PostgreSQL database structure, with necessary validation annotations for the ORM handling and Json output. This change promotes better concurrency handling and scalability in a production environment.

* Set up base.go model for PostgreSQL migration

A Base model was created in the domain to ease the transition from SQLite to PostgreSQL. This new struct aligns with PostgreSQL's database schema, including the necessary ORM and Json validation tags. It will result in improved scalability and concurrent handling in our production environment.

* Add PostgreSQL database connection in db.go

A new file, db.go, was created to establish a connection with a PostgreSQL database. A ConnectDB function was written which attempts to open a connection using a connection string. If the environment variable AutoMigrateDb is set to "true", the database is automatically migrated using Go's gorm.Open functionality. This change is done to improve the scalability of the database operations and handle concurrency better in a production environment.

* Add docker-compose configuration for PostgreSQL

Added a new Docker Compose configuration file to set up a PostgreSQL service. This simplifies the database setup process and ensures that the database runs in a containerized environment, which is easy to scale and manage. Providing settings such as user, password, and database name via environment variables makes the setup flexible and secure.

* Add database configuration in env

Integrated database configuration in environment setup. Added 'DatabaseURL', 'DbType', and 'AutoMigrateDb' to load from environment file. This will provide flexibility to switch between different databases and manage database migrations effectively.

* Add database connection and logging

Added fmt and db imports to main.go for including database connection and logging features. The addition allows the application to print database URL and type, aiding in debugging during development. The db.ConnectDB() function ensures proper connection to the database.

* Add new PR model to the application domain

Introduces a new model 'PR' to the application domain to structure and store information about Pull Requests. Fields in this model include - URL, number, state amongst others. This will allow for a more efficient and organized way to handle PR data in the application.

* Remove unused fmt import and debug print statement

This commit cleans up the code by removing an unused import of the "fmt" package and a debug print statement. This helps in maintaining cleaner and more efficient code.

* Rename database container to DB_NOTIFICATION

Updated the `docker-compose.yml` to rename the database container from `DB_CODE_PIX` to `DB_NOTIFICATION`. This change ensures proper identification and alignment with the new service naming conventions.

* Add Docker setup for Golang app

Include multi-stage Dockerfile and update docker-compose.yml to integrate the Golang app service with PostgreSQL. The app service definition includes port mapping, environment variables, and a dependency on the database service.

* Update database connection to use environment variable

Replaced hardcoded database connection string with `env.DatabaseURL` variable. This enhances security by avoiding hardcoding sensitive credentials in the source code.

* Add `SetEnv` functionality and unit tests

This commit introduces the `SetEnv` variable and its usage across the codebase, including setting environment variables for `GIN_MODE`. It also removes redundant import statements and sets up a new test file for domain conversions. Additionally, it includes some enhancements in the main file for setting trusted proxies and enabling middlewares.